pyproject.toml
[project]
name = "swampyer"
version = "3.0.20240511"
description = "Simple WAMP library with minimal external dependencies"
authors = [{name = "Aki Mimoto", email = "aki+swampyer@zaber.com"}]
requires-python = ">=3.8"
dependencies = [
  "websocket-client>=0.59.0",
  "certifi>=2023.11.17"
]

[project.optional-dependencies]
cbor = [ "cbor" ]
msgpack = [ "msgpack" ]
websockets = [ "websockets" ]
all = [
    "cbor",
    "msgpack",
    "websockets"
]

[tool.pdm]
distribution = true
build = { includes = ["src/swampyer"] }

[build-system]
requires = [
    "pdm-backend"
]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies.test]
pytest = ">=4.6.11"
nox = ">=2023.4.22"
